Added _adsorption_vector attribute to AdsorptionSites
`_single_adsorption` and `_double_adsorption` both call `get_adsorption_vectors` which is relatively slow. When these functions are called for every site on a surface in `add_adsorbates` this gets very slow. The adsorption vectors are the same for every site, so they do not need to be recomputed. This commit creates an attribute `_adsorption_vector` that can be reused for every site.